fix(store): refactor setup of vuex store
#12
