+++ src/store/config.js
@@ -1,0 +1,1 @@
+export const STORAGE_KEY = 'mut-tournament';
