+++ src/store/index.js
@@ -1,12 +1,15 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
-import { state, mutations } from './mutations'
-import plugins from './plugins'
+import Vue from 'vue';
+import Vuex from 'vuex';
+import state from './state';
+import mutations from './mutations';
+import getters from './getters';
+import plugins from './plugins';
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   state,
   mutations,
-  plugins
+  plugins,
+  getters
 })
