+++ src/store/mutations.js
@@ -1,21 +1,20 @@
-export const STORAGE_KEY = 'mut-tournament';
-
-export const state = {
-  tournaments: JSON.parse(window.localStorage.getItem(STORAGE_KEY) || '[]')
-};
-
-export const mutations = {
-  addTournament (state, { name, numberOfPlayers, type }) {
+export default {
+  addTournament(state, { slug, name, numberOfPlayers, type }) {
     state.tournaments.push({
+      slug: slug,
       name,
       numberOfPlayers,
       type
     })
   },
+
+  addTeam(state, { team, tournament }) {
+    const tournament = state.tournaments.find(t => t.slug === tournament);
+
+    if (tournament.name) {
+      if (!tournament.teams) tournament.teams = [];
+      tournament.teams.push(team)
+    }
+  },
 }
 
-export const getters = {
-  numberOfPlayers: state => name => {
-    return state.tournaments.filter(tournament => tournament.name === name).numberOfPlayers
-  }
-}
