+++ src/store/state.js
@@ -1,0 +1,5 @@
+import { STORAGE_KEY } from './config';
+
+export default {
+  tournaments: JSON.parse(window.localStorage.getItem(STORAGE_KEY) || '[]')
+};
